requirements.txt
sqlmodel
pandas
msgraph-sdk
pandas
oracledb
cx_oracle
fastapi
uvicorn
